+++ Spore.h
@@ -36,7 +36,7 @@ public:
     void SporeSpreadDisp(Img& S_umca, Img& S_oaks, Img& I_umca,
                          Img& I_oaks, Img& lvtree_rast, Rtype rtype,
                          double *weather, double weather_value,
-                         double scale1, int kappa = 2,
+                         double scale1, double kappa = 2,
                          Direction wdir = NONE, double scale2 = 0.0,
                          double gamma = 0.0);
     ~Sporulation();
